feat : 디테일 페이지, GNB 레이아웃 조정 #75
